Moving gulpmain, tslint, build_config under tools and gulpfile.js to root folder
