+++ gulpfile.js
@@ -2,7 +2,7 @@
 var fs = require("fs"),
     path = require('path'),
     gulp = require("gulp"),
-    buildConfig = require('./build_config.json');
+    buildConfig = require('./tools/build_config.json');
 
 gulp.on("task_not_found", function (err) {
     console.error("\nPlease run 'npm install' from project root\n");
@@ -12,7 +12,7 @@ gulp.on("task_not_found", function (err) {
  To add additional gulp tasks, add gulpfile in folder and reference it below
  for example:  require('./src/compile/gulpfile');
 */
-var gulpMain = path.join(buildConfig.buildSrc, "gulpmain.js");
+var gulpMain = path.resolve(buildConfig.buildTools, "gulpmain.js");
 if (fs.existsSync(gulpMain)) {
     require(gulpMain);
 }
