--- src/build_config.json
@@ -1,12 +1,0 @@
-{
-    "src": ".",
-    "templates": "../templates",
-    "build": "../build",
-    "buildCoverage": "../build/coverage",
-    "buildSrc": "../build/src",
-    "buildPackages": "../build/packages/node_modules",
-    "buildTools": "../build/tools",
-    "buildTemplates": "../build/packages/node_modules/taco-kits/templates",
-    "tsCompileOptions": { "noImplicitAny": true, "noEmitOnError": true, "target": "ES5", "module": "commonjs", "sourceMap": true, "sortOutput": true }
-
-}
